% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deregister.R
\name{deregister}
\alias{deregister}
\title{Deregister a directory from the index}
\usage{
deregister(dir, url, retry = 3, wait = 1, block = TRUE)
}
\arguments{
\item{dir}{String containing the path to the directory to be deregistered.
This may be a relative or absolute path.}

\item{url}{String containing the URL to the SewerRat REST API.}

\item{retry, wait}{Deprecated arguments, ignored.}

\item{block}{Logical scalar indicating whether to block on successful registration.}
}
\value{
On success, the directory is deregistered and NULL is invisibly returned.

If \code{block=FALSE}, the function returns before confirmation of successful deregistration from the SewerRat API.
This can be useful for asynchronous processing of directories with many files.
}
\description{
Remove a directory on the shared filesystem from the search index.
This assumes that either the directory does not exist,
or that the directory is world-readable and the caller has write access to it.
}
\examples{
# Start up an example SewerRat service:
info <- startSewerRat()

tmp <- tempfile()
dir.create(tmp)
write(file=file.path(tmp, "whee.json"), '{ "foo": "bar" }')
register(tmp, "whee.json", url=info$url)
query("bar", url=info$url)

# After deregistration, the files cannot be queried.
deregister(tmp, url=info$url)
query("bar", url=info$url)
}
\author{
Aaron Lun
}
