% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksave.R
\name{quicksave}
\alias{quicksave}
\title{Quickly save and read objects from disk}
\usage{
quicksave(x, metadata, path)
}
\arguments{
\item{x}{Some Bioconductor object.}

\item{metadata}{A list of metadata describing the object.
Any JSON-compatible data can be stored, in any structure; though it is conventional to have a \code{title} and \code{description} field.}

\item{path}{String containing the path to the directory in which to save \code{x}.
Note that this should be world-readable if it is to be registered in \code{\link{register}}.}
}
\description{
Quickly save and read Bioconductor objects from their on-disk representations (based on the \pkg{alabaster.base} package).
}
\examples{
library(S4Vectors)
df <- DataFrame(A=1:10, B=LETTERS[1:10])
meta <- list(title="A data frame", description="This is a data frame")
tmp <- tempfile()
quicksave(df, meta, tmp)
quickread(tmp)

}
\author{
Aaron Lun
}
