% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register.R
\name{register}
\alias{register}
\title{Register a directory for indexing}
\usage{
register(dir, names, url, retry = 3, wait = 1, block = TRUE)
}
\arguments{
\item{dir}{String containing the path to the directory to be registered.
This may be a relative or absolute path.}

\item{names}{Character vector containing the base names of the metadata files.}

\item{url}{String containing the URL to the SewerRat REST API.}

\item{retry, wait}{Deprecated arguments, ignored.}

\item{block}{Logical scalar indicating whether to block on successful registration.}
}
\value{
On success, the directory is registered and NULL is invisibly returned.
A warning is raised if any particular metadata file cannot be indexed.

If \code{block=FALSE}, the function returns before confirmation of successful registration from the SewerRat API.
This can be useful for asynchronous processing of directories with many files.
}
\description{
Register a directory on a shared filesystem for inclusion into the search index.
This assumes that the directory is world-readable and the caller has write access to it.
}
\examples{
# Start up an example SewerRat service:
info <- startSewerRat()

# Mocking up a directory:
tmp <- tempfile()
dir.create(tmp)
write(file=file.path(tmp, "metadata.json"), '{ "name": "Aaron" }')

# Once we register it, we can query its contents.
register(tmp, "metadata.json", url=info$url)
query("Aaron", url=info$url)
}
\author{
Aaron Lun
}
