% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieveMetadata.R
\name{retrieveMetadata}
\alias{retrieveMetadata}
\title{Retrieve registered metadata}
\usage{
retrieveMetadata(path, url)
}
\arguments{
\item{path}{String containing the absolute path to a file in a registered directory.}

\item{url}{String containing the URL to the SewerRat REST API.}
}
\value{
String containing the path to the file on the caller's filesystem.
}
\description{
Retrieve a single metadata entry in a registered directory from the SewerRat API.
}
\examples{
info <- startSewerRat()

# Mocking up a directory of stuff to query.
mydir <- tempfile()
dir.create(mydir)
write(file=file.path(mydir, "metadata.json"), '{ "first": "Aaron", "last": "Lun" }')
dir.create(file.path(mydir, "diet"))
write(file=file.path(mydir, "diet", "metadata.json"), 
   '{ "meal": "lunch", "ingredients": "water" }')

# Registering it:
register(mydir, "metadata.json", url=info$url)

retrieveMetadata(paste0(mydir, "/metadata.json"), url=info$url)
retrieveMetadata(paste0(mydir, "/diet/metadata.json"), url=info$url)

}
\author{
Aaron Lun
}
