% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startSewerRat.R
\name{startSewerRat}
\alias{startSewerRat}
\alias{stopSewerRat}
\title{Start and stop a SewerRat service}
\usage{
startSewerRat(db = tempfile(fileext = ".sqlite3"), port = NULL, setUrl = TRUE)

stopSewerRat()
}
\arguments{
\item{db}{String containing a path to the SQLite database.}

\item{port}{Integer specifying the port to use for hosting the service.
If \code{NULL}, a free port is randomly selected.}

\item{setUrl}{Logical scalar indicating whether \code{\link{restUrl}} should be updated to use the new service.}
}
\value{
For \code{startSewerRat}, a list indicating whether a new service was set up, along with the port number.
If \code{setUrl=TRUE}, the REST URL is updated to this new service.

For \code{stopSewerRat}, any existing service is shut down, and \code{NULL} is invisibly returned.
The REST URL is also reset to its previous value if it was changed by \code{startSewerRat}.
}
\description{
This sets up a single SewerRat service for the entire R session, is intended for examples and tests.
Real SewerRat deployments should operate outside of R.
}
\examples{
startSewerRat()
restUrl()
stopSewerRat()

}
\seealso{
\url{https://github.com/ArtifactDB/SewerRat}, for source code and binaries to build and run a SewerRat service.
}
