% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deregister.R
\name{deregister}
\alias{deregister}
\title{Deregister a directory from the index}
\usage{
deregister(dir, url, wait = 1)
}
\arguments{
\item{dir}{String containing the path to the directory to be deregistered.
This may be a relative or absolute path.}

\item{url}{String containing the URL to the SewerRat REST API.}

\item{wait}{Numeric scalar specifying the number of seconds to wait for the filesystem to synchronize.}
}
\value{
On success, the directory is deregistered and NULL is invisibly returned.
}
\description{
Remove a directory on the shared filesystem from the search index.
This assumes that either the directory does not exist,
or that the directory is world-readable and the caller has write access to it.
}
\examples{
# Start up an example SewerRat service:
info <- startSewerRat()

tmp <- tempfile()
dir.create(tmp)
write(file=file.path(tmp, "whee.json"), '{ "foo": "bar" }')
register(tmp, "whee.json", url=info$url)
query("bar", url=info$url)

# After deregistration, the files cannot be queried.
deregister(tmp, url=info$url)
query("bar", url=info$url)
}
\author{
Aaron Lun
}
