% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deregister.R
\name{deregister}
\alias{deregister}
\title{Deregister a directory from the index}
\usage{
deregister(dir, url, retry = 3, wait = 1)
}
\arguments{
\item{dir}{String containing the path to the directory to be deregistered.
This may be a relative or absolute path.}

\item{url}{String containing the URL to the SewerRat REST API.}

\item{retry}{Integer scalar specifying the number of times to try to finish the registration.
Larger values may be necessary if \code{dir} is in a network share that takes some time to synchronise.}

\item{wait}{Numeric scalar specifying the number of seconds to wait before each retry.}
}
\value{
On success, the directory is deregistered and NULL is invisibly returned.
}
\description{
Remove a directory on the shared filesystem from the search index.
This assumes that either the directory does not exist,
or that the directory is world-readable and the caller has write access to it.
}
\examples{
# Start up an example SewerRat service:
info <- startSewerRat()

tmp <- tempfile()
dir.create(tmp)
write(file=file.path(tmp, "whee.json"), '{ "foo": "bar" }')
register(tmp, "whee.json", url=info$url)
query("bar", url=info$url)

# After deregistration, the files cannot be queried.
deregister(tmp, url=info$url)
query("bar", url=info$url)
}
\author{
Aaron Lun
}
