% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deregister.R
\name{deregister}
\alias{deregister}
\title{Deregister a directory from the index}
\usage{
deregister(dir, url = restUrl())
}
\arguments{
\item{dir}{String containing the path to the directory to be registered.}

\item{url}{String containing the URL to the SewerRat REST API.}
}
\value{
On success, the directory is registered and NULL is invisibly returned.
A warning is raised if any particular metadata file cannot be indexed.
}
\description{
Remove a directory on the shared filesystem from the search index.
This assumes that either the directory does not exist,
or that the directory is world-readable and you have write access to it.
}
\examples{
# Start up an example SewerRat service:
startSewerRat()

tmp <- tempfile()
dir.create(tmp)
write(file=file.path(tmp, "whee.json"), '{ "foo": "bar" }')
register(tmp, "whee.json")
query("bar")

# After deregistration, the files cannot be queried.
deregister(tmp)
query("bar")

# Okay, stop the service:
stopSewerRat()
}
\author{
Aaron Lun
}
