% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listFiles.R
\name{listFiles}
\alias{listFiles}
\title{List files in a registered directory}
\usage{
listFiles(path, url, recursive = TRUE, forceRemote = FALSE)
}
\arguments{
\item{path}{String containing the path to the registered (sub)directory.}

\item{url}{String containing the URL of the SewerRat REST API.}

\item{recursive}{Logical scalar indicating whether to list files recursively.}

\item{forceRemote}{Logical scalar indicating whether to force remote access,
even if \code{path} is on the same filesystem as the caller.}
}
\value{
Character vector of relative paths of files in \code{path}.
}
\description{
List files in a registered directory or a subdirectory thereof.
This will call the REST API if the caller is not on the same filesystem.
}
\examples{
# Starting up an example SewerRat service:
info <- startSewerRat()

# Mocking up a directory of stuff to query.
mydir <- tempfile()
dir.create(mydir)
write(file=file.path(mydir, "metadata.json"), '{ "first": "Aaron", "last": "Lun" }')
dir.create(file.path(mydir, "diet"))
write(file=file.path(mydir, "diet", "metadata.json"), 
   '{ "meal": "lunch", "ingredients": "water" }')

# Registering it:
register(mydir, "metadata.json", url=info$url)

# List files:
listFiles(mydir, url=info$url)
listFiles(paste0(mydir, "/diet"), url=info$url)

# Forcing remote access.
listFiles(mydir, url=info$url, forceRemote=TRUE)

}
\author{
Aaron Lun
}
