% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register.R
\name{register}
\alias{register}
\title{Register a directory for indexing}
\usage{
register(dir, names, url, wait = 1)
}
\arguments{
\item{dir}{String containing the path to the directory to be registered.}

\item{names}{Character vector containing the base names of the metadata files.}

\item{url}{String containing the URL to the SewerRat REST API.}

\item{wait}{Numeric scalar specifying the number of seconds to wait for the filesystem to synchronize.}
}
\value{
On success, the directory is registered and NULL is invisibly returned.
A warning is raised if any particular metadata file cannot be indexed.
}
\description{
Register a directory on a shared filesystem for inclusion into the search index.
This assumes that the directory is world-readable and the caller has write access to it.
}
\examples{
# Start up an example SewerRat service:
info <- startSewerRat()

# Mocking up a directory:
tmp <- tempfile()
dir.create(tmp)
write(file=file.path(tmp, "metadata.json"), '{ "name": "Aaron" }')

# Once we register it, we can query its contents.
register(tmp, "metadata.json", url=info$url)
query("Aaron", url=info$url)
}
\author{
Aaron Lun
}
