% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register.R
\name{register}
\alias{register}
\title{Register a directory for indexing}
\usage{
register(dir, names, url = restUrl())
}
\arguments{
\item{dir}{String containing the path to the directory to be registered.}

\item{names}{Character vector containing the base names of the metadata files, e.g., \code{metadata.json}.}

\item{url}{String containing the URL to the SewerRat REST API.}
}
\value{
On success, the directory is registered and NULL is invisibly returned.
A warning is raised if any particular metadata file cannot be indexed.
}
\description{
Register a directory on a shared filesystem for inclusion into the search index.
This assumes that the directory is world-readable and you have write access to it.
}
\examples{
if (interactive()) {
     register(paste0("/gstore/user/scratch/", Sys.user()), "metadata.json")
}
}
\author{
Aaron Lun
}
