% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startSewerRat.R
\name{startSewerRat}
\alias{startSewerRat}
\alias{stopSewerRat}
\title{Start and stop a SewerRat service}
\usage{
startSewerRat(
  db = tempfile(fileext = ".sqlite3"),
  port = NULL,
  wait = 1,
  version = "1.2.0",
  overwrite = FALSE
)

stopSewerRat()
}
\arguments{
\item{db}{String containing a path to the SQLite database.}

\item{port}{Integer specifying the port to use for hosting the service.
If \code{NULL}, a free port is randomly selected.}

\item{wait}{Integer specifying the number of seconds to wait for service initialization.}

\item{version}{String containing the desired version of the Gobbler binary.}

\item{overwrite}{Logical scalar indicating whether to redownload the Gobbler binary.}
}
\value{
For \code{startSewerRat}, a list indicating whether a new service was set up, along with the port number and URL to use in other \pkg{SewerRat} functions.

For \code{stopSewerRat}, any existing service is shut down, and \code{NULL} is invisibly returned.
}
\description{
This sets up a single SewerRat service for the entire R session, is intended for examples and tests.
Real SewerRat deployments should operate outside of R.
}
\examples{
startSewerRat()
startSewerRat() # repeated calls just re-use the same instance.

stopSewerRat()
startSewerRat() # initialize a new instance.

}
\seealso{
\url{https://github.com/ArtifactDB/SewerRat}, for source code and binaries to build and run a SewerRat service.
}
